requirements.txt
bittensor>=7
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68
httpx>=0.23.0
python-dotenv>=1.0.0
boto3>=1.26.0